# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 漏洞管理平台 API
  version: 1.0.0
  description: |
    漏洞管理平台对外接口。包含两类调用方：

    - **扫描工具**（机器）：使用 `X-API-Key`，调用 `/ingest/**` 上报漏洞与正负样本。
    - **平台前端**（浏览器）：使用 `Authorization: Bearer <JWT>`，调用查询与处置接口。

    统一响应格式：`{ code, message, data, traceId }`，`code = 0` 表示业务成功。
    时间统一为 ISO 8601 UTC，字段统一小驼峰。
    兼容策略：只增字段，不删不改语义。
  contact:
    name: 平台侧
servers:
  - url: http://localhost:4010/api/v1
    description: 本地 Mock（prism mock openapi.yaml）
  - url: http://localhost:3000/api/v1
    description: 本地开发
  - url: https://vuln.example.com/api/v1
    description: 生产
...
security:
  - bearerAuth: []
paths:
...
  /ingest/scans:
    post:
      tags: [Ingest]
      summary: Step 1 创建扫描批次（幂等，scanNo 为幂等键）
      description: 项目以 (repoType, repoUrl) 唯一匹配，不存在则自动创建。
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateScanRequest' }
      responses:
        '201':
          description: 批次已创建
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/CreateScanResult' }
        '200':
          description: 批次已存在（幂等返回，duplicated=true）
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/CreateScanResult' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/ApiKeyInvalid'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/ServerError'
  /ingest/scans/{scanNo}/vulnerabilities:
    post:
      tags: [Ingest]
      summary: Step 2 批量上报漏洞（1~500 条/批）
      description: 平台按指纹去重。已存在则更新 lastFoundAt；若原状态为 fixed/ignored 则置回 open 并记为 resurfaced。
      security:
        - apiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/ScanNo'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [vulnerabilities]
              properties:
                vulnerabilities:
                  type: array
                  minItems: 1
                  maxItems: 500
                  items: { $ref: '#/components/schemas/VulnerabilityItem' }
      responses:
        '200':
          description: 上报完成（可能部分跳过）
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/VulnIngestResult' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/ApiKeyInvalid'
        '404':
          $ref: '#/components/responses/NotFound'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
  /ingest/scans/{scanNo}/samples:
    post:
      tags: [Ingest]
      summary: Step 3 批量上报正负样本（1~500 条/批）
      description: |
        `label=positive` 表示该文件含漏洞（正样本），`label=negative` 表示无漏洞（负样本）。
        正样本通过 `externalVulnId` 关联**本批次**已上报的漏洞，因此必须先调 Step 2。
      security:
        - apiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/ScanNo'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [samples]
              properties:
                samples:
                  type: array
                  minItems: 1
                  maxItems: 500
                  items: { $ref: '#/components/schemas/SampleItem' }
      responses:
        '200':
          description: 上报完成
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/SampleIngestResult' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/ApiKeyInvalid'
        '404':
          $ref: '#/components/responses/NotFound'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
  /ingest/scans/{scanNo}/complete:
    post:
      tags: [Ingest]
      summary: Step 4 结束扫描批次
      description: 可选但强烈建议调用。未调用时平台定时任务会把超过 2 小时仍为 running 的批次标记为 partial。
      security:
        - apiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/ScanNo'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { $ref: '#/components/schemas/ScanStatus' }
                finishedAt: { type: string, format: date-time, nullable: true }
                totalFiles: { type: integer, minimum: 0 }
                scannedFiles: { type: integer, minimum: 0 }
                errorMessage: { type: string, nullable: true, maxLength: 1000 }
      responses:
        '200':
          description: 批次已结束
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/CompleteScanResult' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/ApiKeyInvalid'
        '404':
          $ref: '#/components/responses/NotFound'
  /ingest/report:
    post:
      tags: [Ingest]
      summary: 一次性全量上报（模式 B，小仓库/调试用）
      description: 等价于 Step 1~4 合并，服务端在单事务内完成。vulnerabilities 与 samples 各 ≤ 500 条。
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FullReportRequest' }
      responses:
        '201':
          description: 上报完成
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/FullReportResult' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/ApiKeyInvalid'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'

  # ------------------------------------------------------- Vulnerabilities
...
components:
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: 扫描工具使用的长期凭证，形如 `vuln_sk_xxxxxxxx`。仅授予 `ingest` 权限范围。
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 平台前端登录后获得，有效期 8 小时。
  parameters:
...
    ScanNo:
      name: scanNo
      in: path
      required: true
      description: 扫描批次号。含 `/` 等特殊字符时必须 URL 编码；建议扫描侧统一用 `-` 分隔。
      schema: { type: string, maxLength: 64 }
...
  responses:
...
    BadRequest:
      description: 参数校验失败
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiResponse' }
          example:
            code: 40001
            message: 参数校验失败
            data:
              errors:
                - field: vulnerabilities[0].severity
                  message: 必须是 critical/high/medium/low/info 之一
            traceId: a1b2c3d4e5f6
...
    ApiKeyInvalid:
      description: API Key 无效、过期或被吊销
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiResponse' }
          example: { code: 40101, message: API Key 无效或已被吊销, data: null }
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiResponse' }
          example: { code: 40400, message: 资源不存在, data: null }
    PayloadTooLarge:
      description: 请求体过大，请使用分片上报接口
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiResponse' }
          example: { code: 41300, message: 请求体过大，请使用分片上报接口, data: null }
    TooManyRequests:
      description: 请求过于频繁，按 Retry-After 头退避重试
      headers:
        Retry-After:
          schema: { type: integer }
          description: 建议等待秒数
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiResponse' }
          example: { code: 42900, message: 请求过于频繁, data: null }
    ServerError:
      description: 服务器内部错误，可指数退避重试
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiResponse' }
          example: { code: 50000, message: 服务器内部错误, data: null }
  schemas:
    ApiResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          description: 0 表示业务成功，非 0 见错误码表
        message: { type: string }
        data: { nullable: true }
        traceId: { type: string }
...
    Severity:
      type: string
      enum: [critical, high, medium, low, info]
      description: 严重 / 高危 / 中危 / 低危 / 提示
...
    ScanStatus:
      type: string
      enum: [running, success, failed, partial]
    RepoType:
      type: string
      enum: [github, gitlab, gitee, bitbucket, other]
    TriggerType:
      type: string
      enum: [push, merge_request, manual, schedule, webhook]
    SampleLabel:
      type: string
      enum: [positive, negative]
      description: positive = 有漏洞（正样本）；negative = 无漏洞（负样本）

    # ---------------- ingest 请求体 ----------------
    ScanObject:
      type: object
      required: [repoType, repoUrl]
      properties:
        repoType: { $ref: '#/components/schemas/RepoType' }
        repoUrl: { type: string, maxLength: 512, example: 'https://github.com/org/demo' }
        repoFullName: { type: string, maxLength: 255, example: 'org/demo' }
        projectName: { type: string, maxLength: 128, example: demo }
        branch: { type: string, maxLength: 128, example: main }
        commitId: { type: string, maxLength: 64 }
        commitMessage: { type: string, maxLength: 512 }
        commitAuthor: { type: string, maxLength: 64 }
        commitTime: { type: string, format: date-time, nullable: true }
    ScannerInfo:
      type: object
      properties:
        name: { type: string, maxLength: 64 }
        version: { type: string, maxLength: 32 }
    VulnerabilityItem:
      type: object
      required: [ruleId, title, severity, filePath]
      properties:
        externalVulnId:
          type: string
          maxLength: 64
          description: 扫描工具内的漏洞 ID，用于与样本关联及双向对账
        ruleId:
          type: string
          maxLength: 128
          description: 规则 ID，参与指纹计算，**必须跨版本稳定**
        ruleName: { type: string, maxLength: 255 }
        title: { type: string, maxLength: 512 }
        severity: { $ref: '#/components/schemas/Severity' }
        category: { type: string, maxLength: 64, example: injection }
        cwe: { type: string, maxLength: 32, example: CWE-89 }
        cve: { type: string, maxLength: 64, nullable: true }
        language: { type: string, maxLength: 32, example: java }
        filePath:
          type: string
          maxLength: 512
          description: 仓库相对路径，统一 `/` 分隔，不带前导 `/`
        lineStart: { type: integer, nullable: true, minimum: 1 }
        lineEnd: { type: integer, nullable: true, minimum: 1 }
        codeSnippet: { type: string, nullable: true, maxLength: 65536 }
        description: { type: string, nullable: true }
        suggestion: { type: string, nullable: true }
        confidence: { type: number, format: float, minimum: 0, maximum: 1, nullable: true }
    SampleItem:
      type: object
      required: [label, filePath]
      properties:
        externalSampleId: { type: string, maxLength: 64 }
        label: { $ref: '#/components/schemas/SampleLabel' }
        filePath: { type: string, maxLength: 512 }
        language: { type: string, maxLength: 32 }
        lineStart: { type: integer, nullable: true }
        lineEnd: { type: integer, nullable: true }
        snippet: { type: string, nullable: true, maxLength: 65536 }
        snippetHash:
          type: string
          maxLength: 64
          description: 片段 sha256（64 位小写十六进制）。不传由平台计算；建议扫描侧提供以加速去重。
        fileHash: { type: string, maxLength: 64, nullable: true }
        externalVulnId:
          type: string
          maxLength: 64
          nullable: true
          description: 仅正样本使用，关联本批次已上报的 VulnerabilityItem.externalVulnId
    CreateScanRequest:
      type: object
      required: [scanNo, scan]
      properties:
        scanNo:
          type: string
          maxLength: 64
          description: 批次号，全局唯一，幂等键。建议 {repoType}-{repoFullName}-{yyyymmdd}-{commitId前7位}，用 `-` 而非 `/`
        scanner: { $ref: '#/components/schemas/ScannerInfo' }
        triggerType: { $ref: '#/components/schemas/TriggerType' }
        scan: { $ref: '#/components/schemas/ScanObject' }
        startedAt: { type: string, format: date-time, nullable: true }
        totalFiles: { type: integer, minimum: 0 }
    FullReportRequest:
      allOf:
        - $ref: '#/components/schemas/CreateScanRequest'
        - type: object
          properties:
            finishedAt: { type: string, format: date-time, nullable: true }
            status: { $ref: '#/components/schemas/ScanStatus' }
            scannedFiles: { type: integer, minimum: 0 }
            vulnerabilities:
              type: array
              maxItems: 500
              items: { $ref: '#/components/schemas/VulnerabilityItem' }
            samples:
              type: array
              maxItems: 500
              items: { $ref: '#/components/schemas/SampleItem' }

    # ---------------- ingest 响应体 ----------------
    CreateScanResult:
      type: object
      properties:
        scanId: { type: integer }
        scanNo: { type: string }
        projectId: { type: integer }
        projectCreated: { type: boolean }
        status: { $ref: '#/components/schemas/ScanStatus' }
        duplicated:
          type: boolean
          description: true 表示该 scanNo 已存在，本次为幂等返回
        createdAt: { type: string, format: date-time }
    VulnIngestResult:
      type: object
      properties:
        received: { type: integer }
        created: { type: integer }
        updated: { type: integer }
        skipped: { type: integer }
        details:
          type: array
          items:
            type: object
            properties:
              externalVulnId: { type: string, nullable: true }
              vulnId: { type: integer }
              vulnNo: { type: string }
              fingerprint: { type: string }
              result:
                type: string
                enum: [created, updated, resurfaced, skipped]
        skippedItems:
          type: array
          items:
            type: object
            properties:
              index: { type: integer }
              externalVulnId: { type: string, nullable: true }
              reason: { type: string }
    SampleIngestResult:
      type: object
      properties:
        received: { type: integer }
        created: { type: integer }
        duplicated: { type: integer }
        truncated: { type: integer, description: 因超过 64KB 被截断的条数 }
        positiveCount: { type: integer }
        negativeCount: { type: integer }
    CompleteScanResult:
      type: object
      properties:
        scanId: { type: integer }
        scanNo: { type: string }
        status: { $ref: '#/components/schemas/ScanStatus' }
        vulnCount: { type: integer }
        sampleCount: { type: integer }
        positiveCount: { type: integer }
        negativeCount: { type: integer }
        durationMs: { type: integer }
    FullReportResult:
      type: object
      properties:
        scanId: { type: integer }
        scanNo: { type: string }
        projectId: { type: integer }
        projectCreated: { type: boolean }
        duplicated: { type: boolean }
        result:
          type: object
          properties:
            vulnReceived: { type: integer }
            vulnCreated: { type: integer }
            vulnUpdated: { type: integer }
            vulnSkipped: { type: integer }
            sampleReceived: { type: integer }
            sampleCreated: { type: integer }
            sampleDuplicated: { type: integer }
            positiveCount: { type: integer }
            negativeCount: { type: integer }

    # ---------------- 查询响应体 ----------------
...